FO-1998: Maltekstvelger skilt ut.
