+++ src/app/veilederverktoy/begrunnelseform/maltekster-form.tsx
@@ -1,55 +1,19 @@
-import React, { ChangeEvent } from 'react';
-import { injectIntl, InjectedIntlProps } from 'react-intl';
+import React from 'react';
 import { BegrunnelseTextArea } from './begrunnelse-textarea';
 import BegrunnelseFooter from './begrunnelse-form-footer';
 import { Form, useFormikContext } from 'formik';
 import { BegrunnelseValues } from './begrunnelse-form';
-import { Select } from 'nav-frontend-skjema';
 import './maltekster-form.less';
+import Maltekstvelger from './maltekstvelger';
 
-interface OwnProps<T extends BegrunnelseValues> {
+interface MalteksterFormProps<T extends BegrunnelseValues> {
     isLoading: boolean;
     maxLength?: number;
     tekstariaLabel: string;
     initialValues: T;
 }
 
-type MalteksterFormProps<T extends BegrunnelseValues> = OwnProps<T> & InjectedIntlProps;
-
 function MalteksterForm(props: MalteksterFormProps<any>) {
-
-    function Maltekstvelger(maltekstvelgerProps: any) {
-
-        function onChange(e: ChangeEvent<HTMLSelectElement>) {
-            if (e.target.value) {
-                const message = props.intl.formatMessage({ id: 'innstillinger.modal.start-eskalering-' + e.target.value});
-                context.setFieldValue('begrunnelse', message);
-                context.setFieldValue('tekst', message);
-            } else {
-                context.resetForm();
-            }
-        }
-
-        const context = useFormikContext();
-        if (maltekstvelgerProps.visible) {
-            return (
-                <Select className="malvelger" onChange={onChange}>
-                    <option value="">Velg en mal</option>
-                    <option value="dagpenger">Dagpenger</option>
-                    <option value="ikke-møtt-møte">Arbeidsavklaringspenger: Ikke møtt til møte</option>
-                    <option value="ikke-deltatt-aktivitet">Arbeidsavklaringspenger: Ikke deltatt på planlagt
-                        aktivitet eller bidrar ikke for å komme i arbeid
-                    </option>
-                    <option value="ikke-deltatt-tiltak">Arbeidsavklaringspenger: Ikke deltatt på tiltak
-                    </option>
-                    <option value="overgangsstønad">Overgangsstønad</option>
-                    <option value="sykepenger">Sykepenger</option>
-                </Select>
-            );
-        } else {
-            return null;
-        }
-    }
 
     const { initialValues } = useFormikContext();
     return (
@@ -63,4 +27,4 @@ function MalteksterForm(props: MalteksterFormProps<any>) {
     );
 }
 
-export default injectIntl(MalteksterForm);
+export default MalteksterForm;
